pyproject.toml
[project]
name = "evaldriven-recsys"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = []

[tool.pytest.ini_options]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "instructor>=1.9.2",
    "langchain-groq>=0.3.5",
    "numpy>=2.3.1",
    "pandas==2.2.3",
    "python-dotenv>=1.1.1",
    "sentence-transformers==3.3.1",
    "weave>=0.51.55",
]
